requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.35.78
botocore==1.35.78
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.2
Django==5.1.4
django-drf-filepond==0.5.0
django-minio-storage==0.5.7
django-otp==1.5.4
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
minio==7.2.12
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.3
s3transfer==0.10.4
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
validate-docbr==1.10.0
